refactor: rename the PickupOptionDialogComponent
