+++ feature-libs/pickup-in-store/components/container/cart-pickup-options-container/cart-pickup-options-container.component.spec.ts
@@ -18,7 +18,7 @@ import { cold } from 'jasmine-marbles';
 import { Observable, of } from 'rxjs';
 import { MockPickupOptionFacade } from '../../../core/facade/pickup-option.service.spec';
 import { PickupOptionsStubComponent } from '../../presentational/pickup-options/pickup-options.component.spec';
-import { MockLaunchDialogService } from '../pickup-delivery-option-dialog/pickup-delivery-option-dialog.component.spec';
+import { MockLaunchDialogService } from '../pickup-option-dialog/pickup-option-dialog.component.spec';
 import {
   CartPickupOptionsContainerComponent,
   cartWithIdAndUserId,
@@ -66,6 +66,7 @@ describe('CartPickupOptionsContainerComponent', () => {
   let activeCartService: ActiveCartFacade;
   let launchDialogService: LaunchDialogService;
   let pickupOptionService: PickupOptionFacade;
+
   const configureTestingModule = () =>
     TestBed.configureTestingModule({
       declarations: [
