+++ feature-libs/pickup-in-store/components/container/index.ts
@@ -6,7 +6,7 @@
 
 export * from './cart-pickup-options-container/index';
 export * from './pdp-pickup-options-container/index';
-export * from './pickup-delivery-option-dialog/index';
+export * from './pickup-delivery-info-container/index';
+export * from './pickup-option-dialog/index';
 export * from './store-list/index';
 export * from './store-search/index';
-export * from './pickup-delivery-info-container/index';
