--- feature-libs/pickup-in-store/components/container/pickup-delivery-option-dialog/index.ts
@@ -1,9 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2022 SAP Spartacus team <spartacus-team@sap.com>
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export * from './pickup-delivery-option-dialog-layout.config';
-export * from './pickup-delivery-option-dialog.component';
-export * from './pickup-delivery-option-dialog.module';
