+++ feature-libs/pickup-in-store/components/container/pickup-option-dialog/index.ts
@@ -1,0 +1,9 @@
+/*
+ * SPDX-FileCopyrightText: 2022 SAP Spartacus team <spartacus-team@sap.com>
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export * from './pickup-option-dialog-layout.config';
+export * from './pickup-option-dialog.component';
+export * from './pickup-option-dialog.module';
