+++ feature-libs/pickup-in-store/components/container/pickup-option-dialog/pickup-option-dialog.component.html
@@ -4,13 +4,13 @@
       <!-- Modal Header -->
       <div class="modal-header cx-dialog-header">
         <div class="cx-dialog-title">
-          {{ 'pickupInStoreDialog.modalHeader' | cxTranslate }}
+          {{ 'pickupOptionDialog.modalHeader' | cxTranslate }}
         </div>
 
         <button
           (click)="close(CLOSE_WITHOUT_SELECTION)"
           class="cx-dialog-close close"
-          [attr.aria-label]="'pickupInStoreDialog.close' | cxTranslate"
+          [attr.aria-label]="'pickupOptionDialog.close' | cxTranslate"
           type="button"
         >
           <span aria-hidden="true">
