+++ feature-libs/pickup-in-store/styles/_index.scss
@@ -1,6 +1,6 @@
-@import './pickup-delivery-option-dialog';
+@import './pickup-delivery-info';
+@import './pickup-option-dialog';
 @import './pickup-options';
-@import './store';
 @import './store-schedule';
 @import './store-search';
-@import './pickup-delivery-info';
+@import './store';
