+++ feature-libs/pickup-in-store/styles/_pickup-option-dialog.scss
@@ -1,7 +1,7 @@
 cx-delivery-pickup-options-dialog {
   background-color: rgba(0, 0, 0, 0.5);
 
-  .cx-delivery-pickup-options-dialog {
+  .cx-pickup-option-dialog {
     @extend .modal-dialog;
     @extend .modal-dialog-centered;
     @include media-breakpoint-down(sm) {
